ficha

+++ BusinessRules/brFichaAtenciones.cs
@@ -10,7 +10,7 @@ namespace BusinessRules
 {
     public class brFichaAtenciones:brGeneral
     {
-        public List<beFichaAtenciones> ListarFichaAtencion(int pID_Ficha)
+        public List<beFichaAtenciones> ListarFichaAtencion(Int64 pID_Ficha)
         {
             List<beFichaAtenciones> lbeAtenciones = null;
             using (SqlConnection con = new SqlConnection(Conexion))
